requirements.txt
django>=4.0
djangorestframework>=3.13
django-filter>=21.1
